Remove tags from taglines
I don't like seeing the tags on the index pages. Might be alright if I
could have tags in the taglines on article pages but not on the index
pages. For now just remove them.

Leave the /tags/ and /tag/foo/ pages in place even though nothing links
to them anymore: I can still browse to them directly or link to them
manually when I want to.